RAC-6490: add fit test cases for ucs-service poller
